Revert "Merge pull request #7 from InspireCampTeam3/DBmigration"
This reverts commit fef4c3119e30b69f9193d7f18ad06353e71c705b, reversing
changes made to 02c704511b9bc7b0eed63c92416a25c4841ffb3f.

+++ src/main/java/com/kitcha/board/publisher/FileEventPublisher.java
@@ -14,7 +14,7 @@ public class FileEventPublisher {
 
     public void sendFileCreateEvent(Board board) {
         Map<String, Object> payload = Map.of(
-                "boardId", board.getBoardId(),
+                "boardId", board.getBoardId().longValue(),
                 "newsTitle", board.getNewsTitle(),
                 "longSummary", board.getLongSummary()
         );
